Reorganised UI module + JSDoc documentation.
