+++ karma.conf.js
@@ -22,7 +22,6 @@ module.exports = function(config) {
       {pattern: 'data/**/*.js'},
       {pattern: 'document/**/*.js'},
       {pattern: 'operator/**/*.js'},
-      {pattern: 'surface/**/*.js'},
       {pattern: 'ui/**/*.js'},
       {pattern: 'node_modules/lodash/**/*.js'},
       {pattern: 'test/fixtures/*.js'},
@@ -38,7 +37,6 @@ module.exports = function(config) {
       'data/**/*.js': ["commonjs", "coverage"],
       'document/**/*.js': ["commonjs", "coverage"],
       'operator/**/*.js': ["commonjs", "coverage"],
-      'surface/**/*.js': ["commonjs", "coverage"],
       'ui/**/*.js': ["commonjs", "coverage"],
       "test/**/*.js": ["commonjs"],
       "node_modules/lodash/**/*.js": ["commonjs"],
