+++ operator/operation.js
@@ -3,12 +3,11 @@
 var Substance = require('../basics');
 
 function Operation() {
+  
 }
 
 Operation.Prototype = function() {
-
   this.isOperation = true;
-
 };
 
 Substance.initClass(Operation);
