+++ ui/surface/index.js
@@ -1,8 +1,15 @@
-var Surface = require('./surface');
-Surface.SurfaceSelection = require('./surface_selection');
+/**
+ * Substance Surface is a generic interface for web-based content editors.
+ * It provides pre-implemented classes for form editors and container editors.
+ *
+ * @module ui/surface
+ */
 
-Surface.FormEditor = require('./form_editor');
-Surface.ContainerEditor = require('./container_editor');
-Surface.Clipboard = require('./clipboard');
+var surface = {};
 
-module.exports = Surface;
+surface.Surface = require('./surface');
+surface.FormEditor = require('./form_editor');
+surface.ContainerEditor = require('./container_editor');
+surface.Clipboard = require('./clipboard');
+
+module.exports = surface;
